tracing: move recordings to new file
Release note: None
